+++ PrimaryHash.h
@@ -1,3 +1,8 @@
+/*
+ * Author: Austin DeLauney
+ *
+ * This class is the meat and potatoes of the program.
+ */
 #ifndef PRIMARYHASH_H
 #define PRIMARYHASH_H
 
@@ -17,6 +22,14 @@ class PrimaryHash
 		//DECONSTRUCTOR
 	   ~PrimaryHash();
 
+	   /*
+	    * printStatistics()
+	    *
+	    * Prints the statistics for this primary hash
+	    *
+	    * Preconditions : Everything is hashed, and good to go
+	    * Postconditions: None
+	    */
 	   void printStatistics();
 
 	   string find(string cityToFind);
@@ -41,7 +54,7 @@ class PrimaryHash
 		 * Precondition : The place is unique
 		 * Postcondition: Returns a hash based on the algorithm used
 		 */ 
-		unsigned long hash(string place);
+		unsigned long long hash(string place);
 
 		/*
 		 * numerizeString(string s)
@@ -51,7 +64,7 @@ class PrimaryHash
 		 * Preconditions : String is unique
 		 * Postconditions: Hash based on sum and constant and stuff
 		 */
-		unsigned long numerizeString(string s);
+		unsigned long long numerizeString(string s);
 
 		unsigned int size, capacity, c, a, b, prime1, prime2;
 		City **m_cities;
